requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Pillow>=10.4.0
PyPDF2==3.0.1
pdfplumber==0.11.4
python-docx==1.1.2
google-generativeai>=0.8.0
requests==2.31.0
